+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Improved
 - fMHA: Fixed out-of-bounds reading for Split-K triton implementation
 - Profiler: fix bug with modules that take a single tuple as argument
+- Profiler: Added manual trigger for a profiling step, by creating a `trigger` file in the profiling directory
 ### Removed
 - Removed support for PyTorch version older than 2.2.0